Extract variables & separate into files under pages
