+++ lib/pages/FooterVariables.js
@@ -1,0 +1,26 @@
+const FooterVariables = {
+    footerSocialLinks: [
+        { 
+            linkClass: "fa-instagram", 
+            linkURL: "https://www.instagram.com/tamudatathon/"
+        },
+        { 
+            linkClass: "fa-linkedin", 
+            linkURL: "https://www.linkedin.com/company/tamudatathon"
+        }, 
+        { 
+            linkClass: "fa-twitter", 
+            linkURL: "https://twitter.com/tamudatathon"
+        },
+        { 
+            linkClass: "fa-facebook", 
+            linkURL: "https://www.facebook.com/tamudatathon/"
+        },
+        { 
+            linkClass: "fa-github", 
+            linkURL: "https://github.com/tamu-datathon-org/"
+        }
+    ]
+};
+
+module.exports = FooterVariables;
