+++ lib/pages/NavbarVariables.js
@@ -1,0 +1,20 @@
+// create all variables for each component
+const NavbarVariables = {
+    navbarSamePageItems: [
+        {
+            name: 'Home',
+            route: '/'
+        }, {
+            name: 'About',
+            route: '#about'
+        }, {
+            name: 'Sponsors',
+            route: '#sponsors'
+        }, {
+            name: 'Team',
+            route: '#team'
+        },
+    ]
+};
+
+module.exports = NavbarVariables;
