+++ lib/pages/SponsorVariables.js
@@ -1,0 +1,138 @@
+const SponsorVariables = {
+    pastSponsors: [
+        {
+            name: "Facebook",
+            img: "/static/img/sponsors/past/facebook.png",
+            url: "https://www.facebook.com/",
+            variant: "xl"
+        },
+        {
+            name: "ConocoPhillips",
+            img: "/static/img/sponsors/past/conocophillips.png",
+            url: "http://www.conocophillips.com/",
+            variant: "lg"
+        },
+        {
+            name: "Aramco",
+            img: "/static/img/sponsors/past/aramco.png",
+            url: "http://www.aramcoservices.com/",
+            variant: "lg"
+        },
+        {
+            name: "Walmart Technology",
+            img: "/static/img/sponsors/past/walmart.png",
+            url: "https://walmart.com/",
+            variant: "lg"
+        },
+        {
+            name: "Motiva",
+            img: "/static/img/sponsors/past/motiva.png",
+            url: "https://motiva.com/",
+            variant: "lg"
+        },
+        {
+            name: "Shell",
+            img: "/static/img/sponsors/past/shell.svg",
+            url: "https://shell.com/",
+            variant: "lg"
+        },
+        {
+            name: "Goldman Sachs",
+            img: "/static/img/sponsors/past/goldmansachs.png",
+            url: "https://goldmansachs.com/",
+            variant: "lg"
+        },
+        {
+            name: "BP",
+            img: "/static/img/sponsors/past/bp.png",
+            url: "https://bp.com/",
+            variant: "lg"
+        },
+        {
+            name: "USAA",
+            img: "/static/img/sponsors/past/usaa.png",
+            url: "https://usaa.com/",
+            variant: "lg"
+        },
+        {
+            name: "MathWorks",
+            img: "/static/img/sponsors/past/mathworks.svg",
+            url: "https://mathworks.com/",
+            variant: "lg"
+        },
+        {
+            name: "CBRE",
+            img: "/static/img/sponsors/past/cbre.png",
+            url: "https://cbre.com/",
+            variant: "lg"
+        },
+        {
+            name: "MLH",
+            img: "/static/img/sponsors/past/mlh.svg",
+            url: "https://mlh.io/",
+            variant: "lg"
+        },
+        {
+            name: "DigitalOcean",
+            img: "/static/img/sponsors/past/digitalocean.png",
+            url: "https://do.co/studenthackathon",
+            variant: "lg"
+        },
+        {
+            name: "Github",
+            img: "/static/img/sponsors/past/github.png",
+            url: "https://digitalocean.com/",
+            variant: "lg"
+        },
+        {
+            name: "Sticker Mule",
+            img: "/static/img/sponsors/past/sticker-mule.png",
+            url: "http://hackp.ac/mlh-stickermule-hackathons",
+            variant: "lg"
+        },
+        {
+            name: "Texas A&M Department of Computer Science and Engineering",
+            img: "/static/img/sponsors/past/csce.png",
+            url: "https://engineering.tamu.edu/cse/index.html",
+            variant: "lg"
+        },
+        {
+            name: "Texas A&M Science Department",
+            img: "/static/img/sponsors/past/science.png",
+            url: "https://www.science.tamu.edu/",
+            variant: "lg"
+        },
+        {
+            name: "Texas A&M Transportation Institute",
+            img: "/static/img/sponsors/past/tti_logo.png",
+            url: "http://tti.tamu.edu/",
+            variant: "lg"
+        },
+        {
+            name: "Texas A&M Institute of Data Science",
+            img: "/static/img/sponsors/past/tamids_logo.png",
+            url: "http://tamids.tamu.edu/",
+            variant: "lg"
+        },
+        {
+            name: "Texas A&M Statistics Department",
+            img: "/static/img/sponsors/past/stat_logo.png",
+            url: "http://stat.tamu.edu/",
+            variant: "lg"
+        },
+        {
+            name: "Texas A&M Math Department",
+            img: "/static/img/sponsors/past/math_logo.png",
+            url: "http://math.tamu.edu/",
+            variant: "lg"
+        },
+        {
+            name: "Texas A&M Data Analytics Club",
+            img: "/static/img/sponsors/past/tdac.png",
+            url: "https://www.tamuanalytics.com/",
+            variant: "lg"
+        },
+    ]
+};
+
+module.exports = SponsorVariables;
